Move SyncContextFactory to SPI module
